MCR-1357 SOLR component depends URN component
Added MCRSolrFileIndexAccumulator to make file indexing more generic.
MyCoRe-URN now has a optional dependency to MyCoRe-SOLR and MyCoRe-SOLR
now does not depends on MyCoRe-URN.
