+++ mycore-solr/src/main/java/org/mycore/solr/index/file/MCRSolrPathDocumentFactory.java
@@ -26,56 +26,71 @@ package org.mycore.solr.index.file;
 import static org.mycore.solr.MCRSolrConstants.CONFIG_PREFIX;
 
 import java.io.IOException;
+import java.lang.reflect.Constructor;
+import java.lang.reflect.InvocationTargetException;
 import java.nio.file.Path;
-import java.nio.file.ProviderMismatchException;
 import java.nio.file.attribute.BasicFileAttributes;
-import java.util.Collection;
-import java.util.Date;
-import java.util.HashSet;
-import java.util.concurrent.TimeUnit;
+import java.util.List;
+import java.util.function.Consumer;
+import java.util.stream.Collectors;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.solr.common.SolrInputDocument;
-import org.mycore.common.MCRCache;
-import org.mycore.common.MCRCache.ModifiedHandle;
 import org.mycore.common.MCRPersistenceException;
 import org.mycore.common.config.MCRConfiguration;
-import org.mycore.datamodel.classifications2.MCRCategLinkReference;
-import org.mycore.datamodel.classifications2.MCRCategLinkServiceFactory;
-import org.mycore.datamodel.classifications2.MCRCategory;
-import org.mycore.datamodel.classifications2.MCRCategoryDAO;
-import org.mycore.datamodel.classifications2.MCRCategoryDAOFactory;
-import org.mycore.datamodel.classifications2.MCRCategoryID;
-import org.mycore.datamodel.common.MCRISO8601Date;
-import org.mycore.datamodel.common.MCRXMLMetadataManager;
-import org.mycore.datamodel.metadata.MCRMetadataManager;
-import org.mycore.datamodel.metadata.MCRObjectID;
-import org.mycore.datamodel.niofs.MCRContentTypes;
+import org.mycore.common.config.MCRConfigurationException;
 import org.mycore.datamodel.niofs.MCRPath;
 import org.mycore.solr.index.handlers.MCRSolrIndexHandlerFactory;
 import org.mycore.solr.index.handlers.stream.MCRSolrFileIndexHandler;
 import org.mycore.solr.index.handlers.stream.MCRSolrFilesIndexHandler;
-import org.mycore.urn.services.MCRURNManager;
-
-import com.google.common.io.Files;
 
 /**
  * @author Thomas Scheffler (yagee)
  */
 public class MCRSolrPathDocumentFactory {
 
+    private static final String ACCUMULATOR_LIST_PROPERTY_NAME = CONFIG_PREFIX + "Indexer.File.AccumulatorList";
+
     private static Logger LOGGER = LogManager.getLogger(MCRSolrPathDocumentFactory.class);
 
     private static MCRSolrPathDocumentFactory instance = MCRConfiguration.instance()
         .<MCRSolrPathDocumentFactory> getInstanceOf(CONFIG_PREFIX + "SolrInputDocument.Path.Factory", (String) null);
 
-    private static final MCRCategoryDAO CATEGORY_DAO = MCRCategoryDAOFactory.getInstance();
+    private static final List<MCRSolrFileIndexAccumulator> ACCUMULATOR_LIST = resolveAccumulators();
 
-    private static final MCRCache<String, String> derivateModified = new MCRCache<>(10000,
-        "derivateID ISODateString cache");
+    /**
+     * @return a list of instances of class listet in {@link #ACCUMULATOR_LIST_PROPERTY_NAME}
+     */
+    private static List<MCRSolrFileIndexAccumulator> resolveAccumulators() {
+        return MCRConfiguration
+            .instance()
+            .getStrings(ACCUMULATOR_LIST_PROPERTY_NAME)
+            .stream()
+            .map((accumulatorClassRef) -> {
+                try {
+                    Class<MCRSolrFileIndexAccumulator> accumulatorClass = (Class<MCRSolrFileIndexAccumulator>) Class
+                        .forName(accumulatorClassRef);
+                    Constructor<MCRSolrFileIndexAccumulator> accumulatorClassConstructor = accumulatorClass
+                        .getConstructor();
+                    MCRSolrFileIndexAccumulator fileIndexAccumulator = accumulatorClassConstructor.newInstance();
+                    return fileIndexAccumulator;
+                } catch (ClassNotFoundException e) {
+                    throw new MCRConfigurationException(
+                        "AccumulatorClass configurated in " + ACCUMULATOR_LIST_PROPERTY_NAME + " not found : "
+                            + accumulatorClassRef, e);
+                } catch (NoSuchMethodException e) {
+                    throw new MCRConfigurationException(
+                        "AccumulatorClass configurated in " + ACCUMULATOR_LIST_PROPERTY_NAME
+                            + " has no default constructor.", e);
+                } catch (IllegalAccessException | InstantiationException | InvocationTargetException e) {
+                    throw new MCRConfigurationException(
+                        "Construxtor of the AccumulatorClass configurated in " + ACCUMULATOR_LIST_PROPERTY_NAME
+                            + " can not be invoked.", e);
 
-    private static MCRXMLMetadataManager XML_MANAGER = MCRXMLMetadataManager.instance();
+                }
+            }).collect(Collectors.toList());
+    }
 
     public static MCRSolrPathDocumentFactory getInstance() {
         return instance;
@@ -91,79 +106,23 @@ public class MCRSolrPathDocumentFactory {
     public SolrInputDocument getDocument(Path input, BasicFileAttributes attr) throws IOException,
         MCRPersistenceException {
         SolrInputDocument doc = new SolrInputDocument();
-        doc.setField("id", input.toUri().toString());
-        String absolutePath = '/' + input.subpath(0, input.getNameCount()).toString();
-        try {
-            MCRPath mcrPath = MCRPath.toMCRPath(input); //check if this is an MCRPath -> more metadata
-            MCRObjectID mcrObjID = MCRMetadataManager.getObjectId(MCRObjectID.getInstance(mcrPath.getOwner()), 10,
-                TimeUnit.SECONDS);
-            if (mcrObjID == null) {
-                LOGGER.warn("Could not determine MCRObject for file " + absolutePath);
-                doc.setField("returnId", mcrPath.getOwner());
-            } else {
-                doc.setField("returnId", mcrObjID.toString());
-                doc.setField("objectProject", mcrObjID.getProjectId());
+
+        Consumer<? super MCRSolrFileIndexAccumulator> accumulate = (accumulator) -> {
+            LOGGER.debug("{} accumulates {}", accumulator, input);
+            try {
+                accumulator.accumulate(doc, input, attr);
+            } catch (IOException e) {
+                LOGGER.error("Error in Accumulator!", e);
             }
-            String ownerID = mcrPath.getOwner();
-            doc.setField("derivateID", ownerID);
-            String urn = MCRURNManager.getURNForFile(ownerID,
-                absolutePath.substring(0, absolutePath.lastIndexOf("/") + 1), input.getFileName().toString());
-            if (urn != null) {
-                doc.setField("fileURN", urn);
-            }
-            doc.setField("derivateModified", getDerivateModified(ownerID));
-            Collection<MCRCategoryID> linksFromReference = MCRCategLinkServiceFactory.getInstance()
-                .getLinksFromReference(new MCRCategLinkReference(mcrPath));
-            HashSet<MCRCategoryID> linkedCategories = new HashSet<>(linksFromReference);
-            for (MCRCategoryID category : linksFromReference) {
-                for (MCRCategory parent : CATEGORY_DAO.getParents(category)) {
-                    linkedCategories.add(parent.getId());
-                }
-            }
-            for (MCRCategoryID category : linkedCategories) {
-                doc.addField("fileCategory", category.toString());
-            }
-        } catch (ProviderMismatchException e) {
-            LOGGER.warn("Cannot build all fields as input is not an instance of MCRPath: " + input);
-        }
-        doc.setField("objectType", "data_file");
-        doc.setField("fileName", input.getFileName().toString());
-        doc.setField("filePath", absolutePath);
-        doc.setField("stream_size", attr.size());
-        doc.setField("stream_name", absolutePath);
-        doc.setField("stream_source_info", input.toString());
-        doc.setField("stream_content_type", MCRContentTypes.probeContentType(input));
-        doc.setField("extension", Files.getFileExtension(input.getFileName().toString()));
-        MCRISO8601Date iDate = new MCRISO8601Date();
-        iDate.setDate(new Date(attr.lastModifiedTime().toMillis()));
-        doc.setField("modified", iDate.getISOString());
+        };
+
+        ACCUMULATOR_LIST.stream().forEach(accumulate);
 
         if (LOGGER.isDebugEnabled()) {
             LOGGER.debug("MCRFile " + input.toString() + " transformed to:\n" + doc.toString());
         }
 
         return doc;
-    }
-
-    /**
-     * returns ISO8601 formated string of when derivate was last modified
-     * 
-     * @param derivateID
-     * @throws IOException
-     *             thrown by {@link ModifiedHandle#getLastModified()}
-     */
-    private static String getDerivateModified(final String derivateID) throws IOException {
-        MCRObjectID derID = MCRObjectID.getInstance(derivateID);
-        ModifiedHandle modifiedHandle = XML_MANAGER.getLastModifiedHandle(derID, 30, TimeUnit.SECONDS);
-        String modified = derivateModified.getIfUpToDate(derivateID, modifiedHandle);
-        if (modified == null) {
-            Date date = new Date(modifiedHandle.getLastModified());
-            MCRISO8601Date date2 = new MCRISO8601Date();
-            date2.setDate(date);
-            modified = date2.getISOString();
-            derivateModified.put(derivateID, modified);
-        }
-        return modified;
     }
 
 }
