+++ mycore-solr/src/main/resources/components/solr/config/mycore.properties
@@ -8,6 +8,7 @@ MCR.EventHandler.MCRPath.100.Class=org.mycore.solr.index.MCRSolrIndexEventHandle
 MCR.URIResolver.ModuleResolver.solr=org.mycore.solr.common.xml.MCRSolrQueryResolver
 MCR.URIResolver.xslImports.solr-document=solr-basetemplate.xsl
 MCR.Module-classbrowser.QueryAdapter=org.mycore.solr.search.MCRSolrQueryAdapter
+MCR.Module-solr.Indexer.File.AccumulatorList=org.mycore.solr.index.file.MCRSolrFileIndexBaseAccumulator
 # define how many mycoreobjects will be sent to solr when rebuilding the metadata index
 MCR.Module-solr.Indexer.BulkSize=100
 MCR.Module-solr.Indexer.ThreadCount=4
