+++ mycore-urn/src/main/resources/components/urn/config/mycore.properties
@@ -6,6 +6,8 @@ MCR.EventHandler.MCRPath.035.Class=org.mycore.urn.events.MCRURNEventHandler
 MCR.Persistence.URN.Store.Class=org.mycore.urn.hibernate.MCRHIBURNStore
 MCR.Hibernate.Mappings=%MCR.Hibernate.Mappings%,org.mycore.urn.hibernate.MCRURN
 
+MCR.Module-solr.Indexer.File.AccumulatorList=%MCR.Module-solr.Indexer.File.AccumulatorList%,org.mycore.urn.indexing.MCRSolrFileIndexURNAccumulator
+
 # XPath to URN in MCRObject metadata
 # default:
 MCR.Persistence.URN.XPath=/mycoreobject/metadata/def.identifier[@class="MCRMetaLangText"]/identifier[@type="urn"]
